+++ src/anomalib/utils/sweep/__init__.py
@@ -4,13 +4,12 @@
 # SPDX-License-Identifier: Apache-2.0
 
 from .config import flatten_sweep_params, get_run_config, set_in_nested_config
-from .helpers import get_openvino_throughput, get_sweep_callbacks, get_torch_throughput
+from .helpers import get_sweep_callbacks, get_torch_throughput
 
 __all__ = [
     "get_run_config",
     "set_in_nested_config",
     "get_sweep_callbacks",
-    "get_openvino_throughput",
     "get_torch_throughput",
     "flatten_sweep_params",
 ]